Fix chipper // Set chipper velocity to 5
