+++ src/joystick/joystick_profiles.h
@@ -18,6 +18,13 @@ const struct joystick_profile profile_default {
     .ROTATION_MULTIPLIER = 1,   // Mutliplier      - Sets rotation speed/multiplier for relative mode
 };
 
+const struct joystick_profile profile_children {
+        .SPEED_MIN = 0.3,           // (0, SPEED_MAX)   - Sets minimum speed. speed < SPEED_MIN = no movement
+        .SPEED_MAX = 1.3,             // (SPEED_MIN, ->)  - Sets maximum speed. speed > SPEED_MAX = SPEED_MAX
+        .DESCRIPTION = "CHILDREN",
+        .ROTATION_MULTIPLIER = 1,   // Mutliplier      - Sets rotation speed/multiplier for relative mode
+};
+
 const struct joystick_profile profile_slow {
     .SPEED_MIN = 0.3,
     .SPEED_MAX = 1.0,
@@ -32,6 +39,6 @@ const struct joystick_profile profile_quick {
     .ROTATION_MULTIPLIER = 1.5,   // Mutliplier      - Sets rotation speed/multiplier for relative mode
 };
 
-#define NUM_JOYSTICK_PROFILES 3
+#define NUM_JOYSTICK_PROFILES 4
 
 #endif //ROBOTEAM_INPUT_JOYSTICK_PROFILES_H
